+++ data-raw/jhu_csse_county_level_subset.R
@@ -1,10 +1,15 @@
+library(readr)
 library(epidatr)
-library(covidcast)
 library(epiprocess)
 library(dplyr)
 
-# Use covidcast::county_census to get the county and state names
-y <- covidcast::county_census %>%
+y <- read_csv("https://github.com/cmu-delphi/covidcast/raw/c89e4d295550ba1540d64d2cc991badf63ad04e5/Python-packages/covidcast-py/covidcast/geo_mappings/county_census.csv", # nolint: line_length_linter
+  col_types = cols(
+    FIPS = col_character(),
+    STNAME = col_character(),
+    CTYNAME = col_character()
+  )
+) %>%
   filter(STNAME %in% c("Massachusetts", "Vermont"), STNAME != CTYNAME) %>%
   select(geo_value = FIPS, county_name = CTYNAME, state_name = STNAME)
 
@@ -18,7 +23,7 @@ jhu_csse_county_level_subset <- pub_covidcast(
   time_values = epirange(20200601, 20211231),
 ) %>%
   select(geo_value, time_value, cases = value) %>%
-  full_join(y, by = "geo_value") %>%
+  inner_join(y, by = "geo_value", relationship = "many-to-one", unmatched = c("error", "drop")) %>%
   as_epi_df()
 
 usethis::use_data(jhu_csse_county_level_subset, overwrite = TRUE)